Add list of reviews to issue data
Review data of pull requests was not yet incorporated in the issue data.
Therefore, a list of reviews is added, extracted from the
/pulls/issue.number/reviews api of GitHub.

Note: The reviews api of GitHub does not only contain actual reviews,
but also treats reply comments to reviews as separate reviews. To
mitigate this, we also extract the corresponding comments from the
/pulls/issue.number/comments api of GitHub and filter the list of
reviews on the following criteria:

A review of the GitHub api is an actual review if...
(a) ... the review does not contain any comments, or
(b) ... the review does contain, at least, one comment that is not a
reply to an existing review comment.

Reviews which do not fulfill one of the above criteria are removed and
are, therefore, not part of the resulting reviews list.

Note: There is a difference between /pulls/issue.number/comments and
/issues/issue.number/comments: The further contains just review comments
(without the initial comment of a review), the latter contains just
comments to the pull request which are not part of a review. To get the
initial comment of a review, /pulls/issue.number/reviews needs to be
considered.

Note: In the resulting json output, the body field for reviews is only
available if hasReviewInitialComment is TRUE, otherwise it will be
omitted as an empty body implies that there is no initial comment.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>