Extract review comments
For each actual review, extract the corresponding comments.

Note: We assgin a comment to the corresponding actual review, that is,
we determine whether a review is a reply and assign the comment to the
review the referenced comment is assigned to.

Note: Extracting all the review comments relies on a correct temporal
order of the comments in the GitHub api. If the referenced comment of a
reply does not exist, the comment will be dropped.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>