Extract GitHub username for commits
When extracting commits (as referenced events), we get the name and
e-mail
from the git commit, but not the GitHub username. As the GitHub username
is available on GitHub for every commit, also store the username for
each
commit. Therefore, the new class 'GitHubCommit' is introduced, which
extends a 'Commit' by adding usernames for author and committer.

In addition, the committer information is added to referenced commit
events,
which was missing previously.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>