Make commits added to a pull request and referenced commits distinguishable
Get committer username also for GitHub commits which were added to a
pull request.

However, if we get a username also for commits added to a pull request,
we cannot differentiate between referenced commits and commits added to
a pull request. The easiest (but not prettiest) solution is to add a
booelan flag to each GitHub commit and set the respective flag when
getting a commit added to a pull request. In addition, this flag also
needs to be extracted when finally serializing the data, that is, such
commits now have the type "commitAddedToPullRequest" instead of
"commit".

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>