Remove 'transient' keyword from the target of ReferencedLink
If the referenced link is a commit, we need to serialize the complete
commit data in order to keep the author and committer data which is
necessary to map usernames to real names and e-mail addresses. To
achieve this, the 'transient' keyword from the target is removed. In the
case that target is not a commit, the target is removed again (as
previously done by 'transient' keyword) while post-serializing. (This is
not a pretty solution, but an appropriate one.) In the case that target
is a commit, the property "target" is renamed to "commit" in the final
output whilel post-serializing.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>