Distinguish related commits
We have two types of related commits (if we neglect commits added to a
pull request):

(1) Commits which are mentioned in a comment of the issue.
(2) Commits which reference the issue in their commit message.

As it makes a difference in the interpretation which of the two types of
commits we have, they now get assigned a different type property.

Why it makes a difference between (1) and (2)? In (1) the issue
references the commit, in (2) the other way round.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>