Store related issues as Integers and not as IssueData objects
To avoid cyclic references, don't store referenced issues as
`ReferencedLink<IssueData>` objects any more, but as
`ReferencedLink<Integer>`. As the resulting output only contains the
issue number in the list of referenced issues, there is no need to store
the complete issue data in the referenced link; it is enough to just
store the number of that issue.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>