Avoid wrongly resolved related issues from cross-repo references
When referencing issues, the `#` sign is used. Up to know, we looked for
all occurrences of `#` followed by a number to determine related issues.
However, this only makes sense if the related issue is in the same
repository. Otherwise, we would wrongly resolve the issue number of
another repository to an issue number inside the analyzed repository.

To avoid such wrong related issues resulting from wrongly parsing
cross-repo references, we need to check for certain patterns. To
reference an issue within the current repository, there are three
different possibilities:

(1) `#number` (e.g., #25)
(2) `user#number` (e.g., se-passau#25)
(3) `user/repo#number` (e.g., se-passau/GitHubWrapper#25)

All the other uses of `#` within a comment may related to an issue in
another repository and are, therefore, neglected when searching for
related issues.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>