Handle missing author and committer of commits
It can be the case that the author or committer field of a commit is
empty in the GitHub API. This needs to be handled to avoid
NullPointerExceptions. In such a case, just a log statement is printed
and no username is set. As the name and e-mail address come from the git
commit itself, those properties should still be available.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>