Fix getter of related issues and add a second setter
To stay consistent with former versions, adjust the getter of related
issues of the `IssueData` class to return a list of actual `IssueData`
objects instead of just a list of `Integer` objects. As related issues
are only stored by the numbers, the corresponding `IssueData` objects
need to be restored from the issue-data cache of the `GitHubRepository`
class.

In addition, to allow setting related issues also from links to
`IssueData` objects, a second setter function
`setRelatedIssuesFromIssueData` is added. This function just takes the
issue number and stores it in the list of links to related issues.

Moreover, fix some outdated documentation.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>