Add workaround for malformed JSON String of commit API
At the moment, the JSON String retrieved from GitHub's commit API
sometimes seems to be malformed, resulting in a JsonSyntaxException.

After quite some debugging, it was not possible to find out why the JSON
String should be malformed. Even when printing the JSON String and
validating it externally, the printed String actually is a valid JSON
String.

However, the problem seems to come from the "files" element of the JSON
String. Maybe there are some ill-encoded characters within the actual
patches of those files, or something similar. As we currently don't
process the "files" element, just clear its content if a
JSONSyntaxException occurs.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>