Fix workaround for malformed JSON String of commit API
The workaround of commit cb9526b25a3cadbb914611445c0bc699da292622 did not work
in all situations: Sometimes, the characters within the "files" element are
that ill-encoded that the regex for the beginning of that element does not
match. Therefore, completely remove all the parts which potentially can
contain ill-encoded characters beforehand. To achieve this, the
'beforeSubstring' method of Apache Commons Lang is used, as this method cuts
the string before reading the ill-encoded characters. However, this introduces
a new dependency (Apache Commons Lang), which is added to the build script.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>