Retry query if an IOException is thrown when crawling data from an URL
If crawling data from an URL does not work for dubious reasons
(resulting in an IOException), don't just ignore the exception and
return an empty Optional, but retry the same query again after waiting
5 seconds.

Notice: This will result in endless retries if the same query
results in an IOException all the time. However, what we have seen so
far was that this exception only occurs randomly (e.g., due to network
connection issues etc.), so endless retries should not occur.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>