Catch NullPointerException when referenced issue is not existent
If a reference to another issue is not existent (for whatever reason),
a NullPointerException was thrown. To circumvent this problem,
the NullPointerException will be catched and no IssueData object for the
referenced issue is returned in this case. That is, such issue
references, where the referenced issue is not existent, are just ignored.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>