Serialize issue data with stream to prevent OutOfMemmoryError
The former implementation led to OutOfMemoryError when serializing a huge amount
of issue data, as the complete JSON String representation of all the issue data
had to be generated and stored in memory before writing it to a file via
BufferedWriter.

To prevent this, use a FileOutputStream together with an OutputStreamWriter to
directly write the JSON String representation of the issue data to file instead
of keeping the whole String in memory.
To achieve this, a new method 'streamSerialize(OutputStreamWriter,
List<IssueData>)' is added to the GitHubRepository class.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>